pyproject.toml
[project]
name = "hand-gesture-recognition"
version = "0.1.0"
description = "Real-time hand gesture recognition using MediaPipe and LSTM neural networks"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE.md"}
authors = [
    {name = "Abdul Ahad"}
]
keywords = ["machine-learning", "computer-vision", "gesture-recognition", "lstm", "mediapipe", "deep-learning"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "kagglehub>=0.3.13",
    "opencv-python>=4.11.0.86",
    "numpy>=1.26.4",
    "pandas>=2.3.3",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "mediapipe>=0.10.21",
    "scikit-learn>=1.7.2",
    "tensorflow>=2.18.1",
    "tqdm>=4.67.1",
    "ipykernel>=7.1.0",
    "ruff>=0.9.0",
    "python-dotenv>=1.2.1",
    "huggingface-hub>=1.0.1",
    "gradio>=5.49.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "jupyter>=1.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
